listening-comp/requirements.txt
boto3>=1.34.0
urllib3<1.27,>=1.25.4
python-dotenv
streamlit
chromadb 